Refactor all github related code to separate files and implement design patterns to improve quality
