+++ src/gurubase-backend/backend/core/utils.py
@@ -1223,7 +1223,7 @@ def ask_question_with_stream(
     user=None,
     github_comments: list | None = None):
     from core.prompts import github_context_template
-    from core.github_handler import GithubAppHandler
+    from core.github.app_handler import GithubAppHandler
 
     start_total = time.perf_counter()
     times = {
@@ -1305,7 +1305,7 @@ def get_summary(question, guru_type, short_answer=False, github_comments: list |
     start_total = time.perf_counter()
     start_prompt_prep = time.perf_counter()
     from core.prompts import summary_template, summary_prompt_widget_addition, summary_prompt_non_widget_addition, github_context_template
-    from core.github_handler import GithubAppHandler
+    from core.github.app_handler import GithubAppHandler
     context_variables = get_guru_type_prompt_map(guru_type)
     context_variables['date'] = datetime.now().strftime("%Y-%m-%d")
     default_settings = get_default_settings()
